redid status effect system
- changed the status effects to use arraylists, so one thing can have multiple effects
- added half of the support tower that debuffs enemies and buffs towers

+++ src/entity/mobs/enemy/HeliBot.java
@@ -140,6 +140,6 @@ public class HeliBot extends Enemy  {
 	}
 	
 	@Override
-	public void damage(int amount) {
+	public void damage(double amount) {
 	}
 }
