+++ src/entity/mobs/pickups/Pickup.java
@@ -39,6 +39,6 @@ public abstract class Pickup extends Mobs{
 		}
 	}
 	@Override
-	public void damage(int amount) {}
+	public void damage(double amount) {}
 
 }
